+++ src/main/java/net/brnbrd/delightful/compat/BrewinChewinCompat.java
@@ -1,7 +1,0 @@
-package net.brnbrd.delightful.compat;
-
-import net.minecraft.resources.ResourceLocation;
-
-public class BrewinChewinCompat {
-	public static final ResourceLocation glowMarmalade = Modid.BC.rl("glow_berry_marmalade");
-}